+++ ltexls-core/src/test/java/org/bsplines/ltexls/parsing/markdown/MarkdownAnnotatedTextBuilderTest.java
@@ -43,6 +43,9 @@ public class MarkdownAnnotatedTextBuilderTest {
         + "multiple lines and [link](example.com)\n",
         "Heading\nParagraph with multiple lines and link\n");
     assertPlainText(
+        "This is a  \r\ntest.\r\n",
+        "This is a test.\n");
+    assertPlainText(
         "# This is a &copy; Test\n"
         + "Another [day &ndash; another](example.com) sentence\n",
         "This is a \u00a9 Test\nAnother day \u2013 another sentence\n");